symblib/dwarf: handle nested subprograms correctly
For some of these we need to emit a new top level declaration, but we
also need to continue to handle inlined programs of the original
subprogram correctly, which is why the nested iter approach doesn't work
I think.

But ideally we'd make the nested iter approach work at some point.